Added small tip on Block\sleep behavior
`Block\sleep` usage can be a little bit confusing, so the loop is not blocked, but it is your program execution.
In terms of non-blocking architecture, what I understand of an sleep is a Promise resolved after N seconds,
From the developer POV, is nice to have this warning here.